pyproject.toml
[tool.poetry]
name = "fight-the-landlord"
version = "0.0.1"
description = ""
authors = ["yanweizhi <yanweizhi@bytedance.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
